% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_networks.R
\name{sccfgl}
\alias{sccfgl}
\title{single-cell Condition-adaptive fused graphical lasso.
The function jointly construct gene co-expression network for multiple class using single-cell Condition-adaptive Fused Graphical Lasso. Pairwise screening matrices are required to adjust between-condition lasso penalty.}
\usage{
sccfgl(
  Y,
  lambda1,
  lambda2,
  btc.screening = NULL,
  rho = 1,
  rho.increment = 1,
  maxiter = 2000,
  tol = 1e-04,
  truncate = 1e-05,
  loglik.trace = FALSE,
  nthre = 10,
  dthre = 0.5,
  ncores = 4,
  change_penalty = FALSE,
  robust_S = TRUE
)
}
\arguments{
\item{Y}{A list of expression data which are n*p matrices. all matrices should have the same p. Values should be 0 or above.}

\item{lambda1}{The tuning parameter for the graphical lasso penalty.}

\item{lambda2}{The tuning parameter for the between condition group lasso penalty.}

\item{btc.screening}{A list of screening matrices (p*p) for between condition penalty. Can be obtained using the function \code{get_scr_mat}.}

\item{rho}{Step size parameter for ADMM algorithm. Large values decrease the step size.}

\item{rho.increment}{Adjustment for rho. In each ADMM iteration, rho will be updated as rho=rho*rho.increment.}

\item{maxiter}{The maximum number of ADMM iterations.}

\item{tol}{The criterion for ADMM convergence.}

\item{truncate}{All value in the estimated inverse convenience below this number will be set to 0.}

\item{loglik.trace}{Store trace of the likelihood of estimation in each iteration.}

\item{nthre}{An integer specifying a threshold on the number of complete observations.}

\item{dthre}{A number specifying the threshold on dropout probabilities.}

\item{ncores:}{set to 1 if using Windows, of >1 can allow for parallelization when calculating the robust correlation}

\item{change_penalty:}{if TRUE, alters the first penalty term to include the multiplicative term}

\item{robust_S:}{if TRUE, uses the robust S instead of the typical estimated covariance}
}
\value{
\code{sccfgl} produces a list that contains estimated precision matrices and other components.
\itemize{
\item{$theta} {The estimation of inverse matrices}
\item{$cor} {The correlation used; Pearson correlation for CFGL and robust correlation otherwise}
\item{$iters} {The number of ADMM iterations}
\item{$loglik.trace} {Trace of log-likelihood}
}
}
\description{
single-cell Condition-adaptive fused graphical lasso.
The function jointly construct gene co-expression network for multiple class using single-cell Condition-adaptive Fused Graphical Lasso. Pairwise screening matrices are required to adjust between-condition lasso penalty.
}
\details{
To run CFGL, set change_penalty=FALSE and robust_S=FALSE. To run CFGL-r, set change_penalty=FALSE and robust_S=TRUE. To run CFGL-rp, set change_penalty=TRUE and robust_S=TRUE. Do not set change_penalty=TRUE and robust_S=FALSE. Note that this function has adaptations from scLink and CFGL.
}
